+++ src/components/layout/SideBar.tsx
@@ -220,11 +220,26 @@ export const pathLists = (isAccounting: boolean, isAdminTask?: boolean) => {
       permission: null,
       subMenu: [
         {
-          path: '/IndexReward',
-          name: 'IndexReward',
-          title: 'นักบินโดรน',
+          path: 'listReward',
+          name: 'listReward',
+          title: 'รายการของรางวัล',
           permission: null,
-          subMenu: [],
+          subMenu: [
+            {
+              path: '/IndexRewardDroner',
+              name: 'IndexRewardDroner',
+              title: 'นักบินโดรน',
+              permission: null,
+              subMenu: [],
+            },
+            {
+              path: '/IndexRewardFarmer',
+              name: 'IndexRewardFarmer',
+              title: 'เกษตรกร',
+              permission: null,
+              subMenu: [],
+            },
+          ],
         },
       ],
     },
@@ -265,12 +280,12 @@ export const pathLists = (isAccounting: boolean, isAdminTask?: boolean) => {
           title: 'รายชื่อผู้ดูแลระบบ',
           subMenu: [],
         },
-        // {
-        //   path: "/IndexPermission",
-        //   name: "IndexPermission",
-        //   title: "บทบาทผู้ดูแล",
-        //   subMenu: [],
-        // },
+        {
+          path: '/IndexPermission',
+          name: 'IndexPermission',
+          title: 'บทบาทผู้ดูแล',
+          subMenu: [],
+        },
       ],
     },
     {