+++ src/pages/task/newTask/ConfirmNewTask.tsx
@@ -84,6 +84,7 @@ export const ConfirmNewTask: React.FC<ConfirmNewTaskProps> = ({
   const handlePaginationChange = (page: number) => {
     setCurrentPage(page)
   }
+
   const startIndex = (currentPage - 1) * pageSize
   const endIndex = startIndex + pageSize
   const currentData = createNewTask?.taskDronerTemp?.slice(startIndex, endIndex)
@@ -408,7 +409,11 @@ export const ConfirmNewTask: React.FC<ConfirmNewTaskProps> = ({
                   <span className='col'>การเตรียมยา</span>
                 </div>
                 <div className='d-flex'>
-                  <span className='col'>{createNewTask?.purposeSprayName}</span>
+                  <span className='col'>
+                    {isEdit
+                      ? createNewTask?.purposeSpray?.purposeSprayName
+                      : createNewTask?.purposeSprayName}
+                  </span>
                   <span className='col'>{createNewTask?.preparationBy}</span>
                 </div>
                 {createNewTask?.preparationRemark && (